--- src/layouts/default/Default.vue
@@ -1,9 +1,0 @@
-<template>
-  <VApp>
-    <DefaultView />
-  </VApp>
-</template>
-
-<script lang="ts" setup>
-import DefaultView from './View.vue'
-</script>
