--- src/layouts/default/View.vue
@@ -1,9 +1,0 @@
-<template>
-  <VMain>
-    <RouterView />
-  </VMain>
-</template>
-
-<script lang="ts" setup>
-//
-</script>
