+++ src/views/App/Admin/Component.vue
