+++ src/views/App/Dashboard/_components/Table/component.vue
