+++ src/views/App/Dashboard/_components/Table/index.ts
