+++ src/views/App/_components/SideNav/SideNavLogo/index.ts
