+++ src/views/App/_components/Toolbar/Component.vue
