+++ src/views/App/_components/index.ts
@@ -1,0 +1,2 @@
+export * from './SideNav'
+export * from './Toolbar'
