+++ src/views/Auth/Login/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Component.vue'
