+++ src/views/Auth/component.vue
@@ -1,0 +1,6 @@
+<script setup lang="ts">
+</script>
+
+<template>
+  <RouterView />
+</template>
