--- src/views/Navigation/SideNav/_components/index.ts
@@ -1,1 +1,0 @@
-export * from './SideNavItem'
