--- src/views/index.ts
@@ -1,6 +1,0 @@
-export * from './Home'
-export * from './Account'
-export * from './Admin'
-export * from './Dashboard'
-export * from './Home'
-export * from './Navigation'
